added conversion from world resoruce to image, exitable trait update -> removed event, added representation unlocking to unlockbuffer and operator, added classiccontext support for simulator, added adding of initial representation and worlds, many bug fixes

+++ source/Simulator/HostSimulatorAuthenticationHandler.h
@@ -196,7 +196,7 @@ namespace Game
 				break;
 			default:
 				Log::Error(L"missing reason case in hostsimulatorauthenticationhandler in onclientregisterfailed",
-					message.content.reason, L"reason");
+					(int) message.content.reason, L"reason");
 
 				sendAuthenticationRejectedMessage(
 					::Net::Host::AuthenticationRejectedMessageContent::Reason::OperatorRequestFailed
